license-monitor/requirements.txt
python-dotenv==1.0.0  # Environment variable management
pydantic==2.5.0  # Data validation and settings management
pydantic-settings==2.1.0  # Settings management for Pydantic
pyyaml==6.0.1  # YAML configuration file parsing
sqlalchemy==2.0.23  # Database ORM for storing license data
pandas==2.1.3  # Data analysis and manipulation
numpy==1.26.2  # Numerical operations
jinja2==3.1.2  # Template engine for report generation
openpyxl==3.1.2  # Excel file generation for reports
requests==2.31.0  # HTTP library for API calls
python-dateutil==2.8.2  # Date and time utilities
pytest==7.4.3  # Testing framework
pytest-cov==4.1.0  # Test coverage reporting
pytest-mock==3.12.0  # Mocking for tests
